Fixed an issue where VideoPlayerControls were hidden when the mouse was moved outside of the video
